+++ templates/boilerplate-template/scripts/move/publish.js
@@ -12,7 +12,7 @@ async function publish() {
 
   move
     .createObjectAndPublishPackage({
-      packageDirectoryPath: "move",
+      packageDirectoryPath: "contract",
       addressName: "message_board_addr",
       namedAddresses: {
         // Publish module to new object, but since we create the object on the fly, we fill in the publisher's account address here