+++ templates/token-minting-dapp-template/scripts/move/compile.js
@@ -28,7 +28,7 @@ async function compile() {
   const move = new cli.Move();
 
   await move.compile({
-    packageDirectoryPath: "move",
+    packageDirectoryPath: "contract",
     namedAddresses: {
       // Publish module to account address
       launchpad_addr: accountAddress,